Reduce GC-churn on the redis subscriptions data in Streaming
Previously we used an object from which we deleted keys (channels) from when no longer needed, and the repeated mutations to that object could incur a GC cost. By switching to a Map, we can skip the costs associated with `delete` on an object.

Ultimately though, we shouldn't be storing callbacks in this anyway, and just be storing connection identifiers, which would make the code much easier to follow and more memory efficent.